requirements.txt
certifi==2020.6.20
cycler==0.10.0
decorator==4.4.2
imageio==2.8.0
joblib==0.15.1
kiwisolver==1.2.0
matplotlib==3.2.2
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
networkx==2.4
numpy==1.18.5
olefile==0.46
pandas==1.0.5
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.5.0
six==1.15.0
sklearn==0.0
sweetviz==1.0a7
threadpoolctl==2.1.0
tifffile==2020.6.3
torch==1.5.1
torchvision==0.6.1
tqdm==4.46.1
wincertstore==0.2
